backend/tweet_results/requirements.txt
python-twitter
amqplib
